% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-peaks.R
\name{.detect_peaks_gradient}
\alias{.detect_peaks_gradient}
\title{Detect SCR peaks using first-derivative zero-crossing}
\usage{
.detect_peaks_gradient(sig, sr, amplitude_min, rise_time_min, rise_time_max)
}
\description{
Detect SCR peaks using first-derivative zero-crossing
}
\keyword{internal}
