% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PhysioEDA-package}
\alias{PhysioEDA}
\alias{PhysioEDA-package}
\title{PhysioEDA: Electrodermal Activity Analysis for PhysioExperiment Objects}
\description{
Provides electrodermal activity (EDA / skin conductance) analysis functions for PhysioExperiment objects. Includes preprocessing (filtering, downsampling), artifact detection and correction, tonic/phasic decomposition (highpass, median, CDA, cvxEDA), SCR peak detection, feature extraction, data transformations, signal quality assessment, visualization, and event-related SCR analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioEDA}
  \item \url{https://x-biosignal.r-universe.dev/PhysioEDA}
  \item \url{https://x-biosignal.github.io/PhysioEDA/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioEDA/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
